1st solution
Failed for some test cases. There are flaw in my algorithm, my assumption is sum left is sum value in left pivot candidate and right sum is sum value in right pivot candidate. So the pivot candidate between nums[1] and nums[nums.lenght()-2]

Until I hit case [-1,-1,-1,0,1,1]
Answer for case is [0], because sum right is 0, then sum left is 0. Why sum left is 0? I think this is a assumption from the creator. In real interview, we must always ask for assumption from the interviewer!