3rd solution
This solution is from the Solution tab.
The idea is we sum all the value. Assuming leftSum is 0. Then from 1st index we compare leftSum with sum - leftSum - nums[i].
leftSum == (sum - leftSum - nums[i]) , this part means we decrease the sum with leftSum + current index loop. If false we add current index to leftSum.
The idea is after we know total value, we decrease the value from the left untuk leftSum == rightSum.