1st solution
My idea is move one by one element from input to result then add the last element by one.
But this solution is wrong. There are condition that I miss understood, "each element in the array contains a single digit". Example if the last digit is 9, the new element will be [1,0]. So I can't assume the length or input array and result array same.